+++ packages/sdk/src/index.ts
@@ -16,6 +16,7 @@ export * from './observable/persistedObservable'
 export * from './persistenceStore'
 export * from './riverConfig'
 export * from './riverDbManager'
+export * from './rpcInterceptors'
 export * from './sign'
 export * from './signerContext'
 export * from './store/store'
